+++ src/main/java/dev/ftb/packcompanion/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package dev.ftb.packcompanion;
+
+import javax.annotation.ParametersAreNonnullByDefault;